+++ .storybook/ShareQRCode.js
@@ -3,7 +3,7 @@ import {storiesOf} from '@storybook/react'
 
 import {RequestQRCode, Action} from '../index'
 import {Logo} from './Logo'
-import {Manager} from './Manager'
+import {Updating} from './Updating'
 
 const defaultData = {
   action: Action.attestation,
@@ -21,4 +21,4 @@ storiesOf('RequestQRCode', module)
   .add('Colors', () => <RequestQRCode requestData={defaultData} bgColor={'#EBF0F1'} fgColor={'#3C3C3D'} />)
   .add('Logo', () => <Logo requestData={defaultData} />)
   .add('Size', () => <RequestQRCode requestData={defaultData} size={300} />)
-  .add('Managed', () => <Manager requestData={defaultData} />)
+  .add('Updating', () => <Updating />)