+++ index.ts
@@ -1,3 +1,4 @@
+import {generateRequestQRCode} from './src/generateRequestQRCode'
 import {createRequestQRCode, updateRequestQRCode, removeRequestQRCode} from './src/manageRequestQRCode'
 import {RequestQRCode} from './src/RequestQRCode'
 import {Action, RequestData, Options, NonceData, Nonces, ResponseData} from './src/types'
@@ -8,6 +9,7 @@ export {
   createRequestQRCode,
   updateRequestQRCode,
   removeRequestQRCode,
+  generateRequestQRCode,
   RequestQRCode,
   Action,
   RequestData,